+++ pages/user/tabs/activity.php
@@ -11,7 +11,6 @@ if ( ! defined( 'ABSPATH' ) ) {
 // Get current user iD
 $user_id = get_current_user_id();
 
-/*
 // Get profile economy
 $profile_economy = loopis_ledger_economy($user_id);
 $payments_membership = $profile_economy['payments_membership'];
@@ -27,22 +26,6 @@ $star_coins = $profile_economy['star_coins'];
 $clovers = $profile_economy['clovers'];
 $clover_coins = $profile_economy['clover_coins'];
 $coins = $profile_economy['coins'];
-*/
-
-// Mock data for now
-$payments_membership = 0;
-$payments_coins = 0;
-$membership_coins = 0;
-$bought_coins = 0;
-$count_given = 0;
-$count_booked = 0;
-$count_submitted = 0;
-$count_deleted = 0;
-$stars = 0;
-$star_coins = 0;
-$clovers = 0;
-$clover_coins = 0;
-$coins = 0;
 ?>
 
 <h3>🧮 Min aktivitet</h3>
@@ -51,14 +34,14 @@ $coins = 0;
 
 <!--COINS-->
 <div class="economy wrapped">
-<p>Regnbågsmynt<span class="right"><img src="<?php echo LOOPIS_THEME_HQ_URI; ?>/assets/img/coin.png" alt="coin symbol" style="height:15px; width: auto;"></span></p>
+<p>Regnbågsmynt<span class="right"><img src="<?php echo LOOPIS_THEME_URI; ?>/assets/img/coin.png" alt="coin symbol" style="height:15px; width: auto;"></span></p>
 <hr>
 <p><b><?php echo $payments_membership; ?></b> köp av medlemskap <span class="plus right">+<?php echo $membership_coins; ?></span></p>
 <?php if ( $payments_coins > 0 ) { ?>
 <p><b><?php echo $payments_coins; ?></b> köp av extra mynt <span class="plus right">+<?php echo $bought_coins; ?></span></p>
 <?php } ?>
 <p><b><?php echo $count_given; ?></b> saker lämnade <span class="plus right">+<?php echo $count_given; ?></span></p>
-<p><b><?php echo $count_booked; ?></b> saker hämtade/paxade <span class="minus right">–<?php echo $count_booked; ?></span></p>
+<p><b><?php echo $count_booked; ?></b> saker paxade <span class="minus right">–<?php echo $count_booked; ?></span></p>
 <hr>
 <p>&nbsp;<span class="right">Totalt: <b><?php echo $coins - $clover_coins - $star_coins; ?></b></span></p>
 </div>
@@ -84,7 +67,7 @@ $coins = 0;
 <div class="economy wrapped">
 <p>Guldstjärnor<span class="right">🌟</span></p>
 <hr>
-<?php include_once LOOPIS_THEME_HQ_DIR . '/templates/user/profile/user-rewards.php'; ?>
+<?php include LOOPIS_THEME_DIR . '/includes/output/user-data/user-rewards.php'; ?>
 <hr>
 <p>&nbsp;<span class="right">Totalt: <b><?php echo $stars; ?></b></span></p>
 <!--Reward-->
@@ -107,10 +90,10 @@ $coins = 0;
 </p>
 
 <!--FAQ-->	
-<p><span class="link"><a href="<?php echo get_permalink( get_page_by_path('hur-funkar-beloningar') ); ?>">📌 Hur funkar belöningar?</a></span></p>
+<p><span class="link"><a href="<?php echo esc_url(network_home_url( '/faq/hur-funkar-beloningar' )); ?>">📌 Hur funkar belöningar?</a></span></p>
 
 <!--Payments-->	
 <h3>📒 Mina kvitton</h3>
 <hr>
 <p>Dina registrerade betalningar till föreningen.</p>
-<?php include_once LOOPIS_THEME_HQ_DIR . '/templates/user/profile/user-payments.php'; ?>
+<?php include_once LOOPIS_THEME_DIR . '/includes/output/user-data/user-payments.php'; ?>