+++ templates/develooper/test-coins.php
@@ -12,9 +12,12 @@ if (!defined('ABSPATH')) {
 
 <div class="admin-block">
     <?php include_once LOOPIS_THEME_HQ_DIR . '/templates/links/developer-link.php' ?>
-    <h5>⚠ Testläge</h5>
+    <h5>⚠ Sidan är i testläge</h5>
     <hr>
-    <p class="small">💡 Använd kortnummer "4242 4242 4242 4242" för att simulera en betalning med Stripe sandbox.</p>
+    <?php if (defined('LOOPIS_STRIPE_WEBHOOK_SECRET_COINS')) { echo LOOPIS_STRIPE_WEBHOOK_SECRET_COINS; } ?>
+    
+    <p class="small">💳 Använd kortnummer "4242 4242 4242 4242" för att simulera en betalning med Stripe sandbox.</p>
+    <p class="small">💡 Du kan också skippa betalning och använda knappen nedan.</p>
 
     <?php if(isset($_POST['activate'])) { add_coins(); refresh_page(); } ?>
 		<form method="post" class="arb" action=""><button name="activate" type="submit" class="purple small">Ge mig 5 mynt</button></form>