+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # 📋 Changelog for "LOOPIS Theme HQ"
 
+## 1.01 (2026-06-22)
+- Maintenance mode revised
+- Pages content for /user and /shop moved from "LOOPIS Theme" to here
+- Script assets (js) centralized and moved to "LOOPIS Theme"
+- Stripe fixes in process...
+
+## 1.00 (2026-06-16)
+- Deployment on new multisite!
+
 ## 0.07 (2026-06-15)
 - Theme constants definition moved to new mu-plugin "LOOPIS Constants"
 - Content of pages /shop and /user are now fetched from "LOOPIS Theme"